Split themed cookie banner by presentation
The themed banner was one component that built both the banner card and
the preference panel, then mutated the DOM at runtime and leaned on
server-remapped text keys to pick wording. Each new rule piled more
branching into that single component and the server remap.

Turn the themed banner into a thin dispatcher that mounts a focused
renderer per presentation (opt-in, opt-out, notice), each owning only its
own markup, wording keys, and behavior. Drive button visibility and
banner validation from the structured layout instead of text presence,
and add a typed per-variant text resolver clients can use directly.

Isolate the SDK-version compatibility logic in compatibility.go and
rename isLegacySDK to the positive capability supportsTextRemap. The
server text remap is now a shim for pre-0.11 clients only; layout-aware
clients receive the raw keys and select their own wording.
