+++ packages/cookie-banner/CHANGELOG.md
@@ -10,12 +10,15 @@ All notable changes to the `@probo/cookie-banner` SDK will be documented in this
 - Soft-validate that `<probo-settings-link>` is present (warn + `probo-validation`), since it is the sole reopen control
 - Structured `layout` object in the banner config (`presentation`, `initial_state`, `reopen_state`, `default_non_necessary_granted`, `buttons`, `settings_link`, `text_variant`). The banner now renders from these explicit fields instead of inferring behavior from hardcoded regulation checks and empty text keys. A config without `layout` means a self-hosted Probo backend older than this SDK: the client logs an error asking to update probod and falls back to a strict opt-in layout
 - First-class notice-only presentation (single acknowledge button) for jurisdictions with no cookie-consent law and implied-consent regimes (Japan APPI, Mexico LFPDPPP)
+- The themed `<probo-cookie-banner>` is now a thin dispatcher that mounts a focused renderer per presentation (opt-in / opt-out / notice); each renderer owns only its own markup, wording keys, and quirks, replacing the single component's runtime button-toggling
+- Export `resolveLayout` and `resolveBannerText` (plus the `BannerLayout` / `BannerText` / `Presentation` types) so headless integrators can resolve the active presentation and its wording
 
 ### Changed
 
 - Under CCPA, the banner starts closed by default (cookies already follow the opt-out model until the visitor opts out)
 - Auto-filled settings-link text inherits font and color from the host; the CCPA icon scales with `1em`
 - Initial state, reopen target, default toggle state, button visibility, and the statutory settings-link style are now driven by the server-provided `layout` rather than `regulation === "CCPA"` / `consent_mode` branching
+- Layout-aware clients (0.11+) receive the raw text keys and select their own wording per `layout.text_variant`; the server-side key remap is now a shim for pre-0.11 clients only. Headless integrators on 0.11+ should use `resolveBannerText` / `layout` to render the correct copy
 
 ### Removed
