+++ packages/cookie-banner/src/layout.ts
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-import type { BannerConfig, BannerLayout } from "./types";
+import type { BannerConfig, BannerLayout, BannerText, TextVariant } from "./types";
 
 // Strict, GDPR-safe layout used only as a defensive fallback: a config without a
 // `layout` means the (self-hosted) Probo backend is older than this SDK. We keep
@@ -53,4 +53,58 @@ export function resolveLayout(config: BannerConfig): BannerLayout {
   }
 
   return STRICT_LAYOUT;
+}
+
+// The banner-card text keys each presentation variant uses. Every variant falls
+// back to the base keys when a customer translation omits the variant-specific
+// wording.
+const VARIANT_TEXT_KEYS: Record<
+  TextVariant,
+  { title: string; description: string; primary: string; secondary?: string }
+> = {
+  default: {
+    title: "banner_title",
+    description: "banner_description",
+    primary: "button_accept_all",
+    secondary: "button_reject_all",
+  },
+  opt_out: {
+    title: "banner_title_opt_out",
+    description: "banner_description_opt_out",
+    primary: "button_acknowledge",
+    secondary: "button_opt_out",
+  },
+  notice: {
+    title: "banner_title_notice",
+    description: "banner_description_notice",
+    primary: "button_dismiss",
+  },
+};
+
+// resolveBannerText builds the typed banner-card wording for the active
+// presentation from the flat text keys, so callers never touch variant-specific
+// key names. The themed renderers use their own keys directly; this is the
+// resolver headless integrators use to render the correct copy.
+export function resolveBannerText(config: BannerConfig): BannerText {
+  const variant = resolveLayout(config).text_variant;
+  const keys = VARIANT_TEXT_KEYS[variant] ?? VARIANT_TEXT_KEYS.default;
+  const texts = config.texts ?? {};
+
+  const pick = (variantKey: string, baseKey: string): string =>
+    texts[variantKey] || texts[baseKey] || "";
+
+  const result: BannerText = {
+    title: pick(keys.title, "banner_title"),
+    description: pick(keys.description, "banner_description"),
+    primaryButton: pick(keys.primary, "button_accept_all"),
+  };
+
+  if (keys.secondary) {
+    const secondary = pick(keys.secondary, "button_reject_all");
+    if (secondary) {
+      result.secondaryButton = secondary;
+    }
+  }
+
+  return result;
 }
